Restored grayscale signal
