+++ res/hash.txt
@@ -1,1 +1,1 @@
-654f9bf
+34e0ae7
