+++ src/main_window.py
@@ -1,10 +1,9 @@
 import json
 import pickle
 from multiprocessing import Manager, Process, Queue, shared_memory
-from multiprocessing.managers import SharedMemoryManager
 
 import numpy as np
-from PySide6.QtCore import QObject, Qt, QTimer, Signal
+from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QApplication, QHBoxLayout, QSplitter, QWidget
 from qframelesswindow import FramelessMainWindow
 
@@ -13,140 +12,9 @@ from daemon import Daemon
 from helpers.image_conversion import numpy_to_pixmap, qimage_to_numpy
 from helpers.paths import config_path
 from preferences import PreferencesDialog
+from queue_io import QueueReader, QueueWriter
 from sidebar import SideBar
 from viewer import PhotoViewer
-
-
-class QueueReader(QObject):
-    received_array = Signal(str, tuple)
-    rotated = Signal(tuple)
-    show_processing_label = Signal(bool)
-    close_shm = Signal()
-    received_processed = Signal(str, bool)
-    received_notification = Signal(str, int)
-
-    def __init__(self, queue, window=None, interval=50):
-        super().__init__()
-        self.queue = queue
-        self.window = window
-        self.timer = QTimer()
-        self.timer.timeout.connect(self.check_queue)
-        self.timer.start(interval)
-
-    def check_queue(self):
-        while not self.queue.empty():
-            message = self.queue.get()
-            print(message)
-            if message["type"] == "shared_array":
-                name = message["name"]
-                size = message["size"]
-                self.received_array.emit(name, size)
-            elif message["type"] == "close_shm":
-                self.close_shm.emit()
-            elif message["type"] == "display_image":
-                array = message["array"]
-                reset = message["reset"]
-                self.received_processed.emit(array, reset)
-            elif message["type"] == "started_processing":
-                self.show_processing_label.emit(True)
-            elif message["type"] == "notification":
-                notification = message["notification"]
-                duration = message["duration"]
-                self.received_notification.emit(notification, duration)
-            elif message["type"] == "enable_toolbox":
-                state = message["state"]
-                self.window.sidebar.toolbox.enable_buttons(state)
-
-            # self.message_received.emit(message)
-
-
-class QueueWriter(QObject):
-    rotate = Signal(bool)
-
-    def __init__(self, queue, window=None):
-        super().__init__()
-        self.queue = queue
-        self.window = window
-
-    def close(self):
-        message = {"type": "exit"}
-        self.queue.put(message)
-
-    def load_image(self, path):
-        message = {"type": "load_image", "path": path}
-        self.queue.put(message)
-        self.window.display_processing_label(True)
-
-    def load_from_clipboard(self):
-        message = {"type": "load_from_clipboard"}
-        self.queue.put(message)
-        self.window.display_processing_label(True)
-
-    def send_pickled_image(self, pickled_data):
-        message = {"type": "load_from_pickle", "data": pickled_data}
-        self.queue.put(message)
-        self.window.display_processing_label(True)
-
-    def send_url(self, url):
-        message = {"type": "load_from_url", "url": url}
-        self.queue.put(message)
-
-    def save_image(self):
-        message = {"type": "save_image"}
-        self.queue.put(message)
-
-    def save_to_clipboard(self):
-        message = {"type": "save_to_clipboard"}
-        self.queue.put(message)
-
-    def reset(self):
-        message = {"type": "reset_storage"}
-        self.queue.put(message)
-        self.window.reset_viewer()
-
-    def send_rotate(self, cw):
-        message = {"type": "rotate", "cw": cw}
-        self.queue.put(message)
-        self.rotate.emit(cw)
-
-    def send_flip(self):
-        message = {"type": "flip"}
-        self.queue.put(message)
-
-    def send_invert(self):
-        message = {"type": "invert"}
-        self.queue.put(message)
-
-    def send_color(self, color, swatch):
-        message = {"type": "change_color", "color": color, "swatch": swatch}
-        self.queue.put(message)
-
-    def save_like_preview(self, value):
-        message = {"type": "save_like_preview", "value": value}
-        self.queue.put(message)
-
-    def send_halftone(self, algorithm, settings):
-        message = {
-            "type": "halftone_settings",
-            "algorithm": algorithm,
-            "settings": settings,
-        }
-        self.queue.put(message)
-
-    def send_grayscale(self, mode, settings):
-        message = {
-            "type": "grayscale_settings",
-            "mode": mode,
-            "settings": settings,
-        }
-        self.queue.put(message)
-
-    def send_enhance(self, settings):
-        message = {
-            "type": "enhance_settings",
-            "settings": settings,
-        }
-        self.queue.put(message)
 
 
 class MainWindow(FramelessMainWindow):
@@ -164,7 +32,6 @@ class MainWindow(FramelessMainWindow):
         self.res_queue = Queue()
         # manager for simple values
         self.manager = Manager()
-        self.smm = SharedMemoryManager()
 
         self.paths = self.manager.dict()
 
@@ -183,6 +50,7 @@ class MainWindow(FramelessMainWindow):
 
         self.reader = QueueReader(self.res_queue, window=self)
 
+        # READER SIGNALS
         self.reader.received_array.connect(self.init_array)
         self.reader.close_shm.connect(self.close_shm)
         self.reader.received_processed.connect(self.display_processed_image)
@@ -190,6 +58,8 @@ class MainWindow(FramelessMainWindow):
         self.reader.show_processing_label.connect(self.display_processing_label)
 
         self.writer = QueueWriter(self.req_queue, window=self)
+
+        # WRITER SIGNALS
         self.writer.rotate.connect(self.rotate_shm)
 
     def _setup_ui(self):
